add: split to_rust func
