+++ src/to_ts_rust/to_rust/func/mod.rs
@@ -1,0 +1,1 @@
+pub mod func;
