+++ src/to_ts_rust/to_rust/mod.rs
@@ -2,3 +2,4 @@ pub mod to_rust;
 mod common_methods;
 mod variable_declaration;
 mod expr_stmt;
+mod func;
